+++ api/v1alpha1/applyconfiguration/utils.go
@@ -32,6 +32,8 @@ func ForKind(kind schema.GroupVersionKind) interface{} {
 		return &apiv1alpha1.ArtifactInputsApplyConfiguration{}
 	case v1alpha1.SchemeGroupVersion.WithKind("ArtifactOutboxStatus"):
 		return &apiv1alpha1.ArtifactOutboxStatusApplyConfiguration{}
+	case v1alpha1.SchemeGroupVersion.WithKind("ArtifactRecordingMetadata"):
+		return &apiv1alpha1.ArtifactRecordingMetadataApplyConfiguration{}
 	case v1alpha1.SchemeGroupVersion.WithKind("ArtifactResourceReference"):
 		return &apiv1alpha1.ArtifactResourceReferenceApplyConfiguration{}
 	case v1alpha1.SchemeGroupVersion.WithKind("ArtifactSessionReference"):
@@ -126,6 +128,8 @@ func ForKind(kind schema.GroupVersionKind) interface{} {
 		return &apiv1alpha1.DebugSessionApproversApplyConfiguration{}
 	case v1alpha1.SchemeGroupVersion.WithKind("DebugSessionArtifact"):
 		return &apiv1alpha1.DebugSessionArtifactApplyConfiguration{}
+	case v1alpha1.SchemeGroupVersion.WithKind("DebugSessionArtifactCollection"):
+		return &apiv1alpha1.DebugSessionArtifactCollectionApplyConfiguration{}
 	case v1alpha1.SchemeGroupVersion.WithKind("DebugSessionArtifactSpec"):
 		return &apiv1alpha1.DebugSessionArtifactSpecApplyConfiguration{}
 	case v1alpha1.SchemeGroupVersion.WithKind("DebugSessionArtifactStatus"):